Add deletion of empty product-groups and style improvements
The deletion is only possible if there are no more children.  It is
there via a separate button.  We should probably do something to
ensure users are sure they want to delete the item, though in practice
it's okay as limited information can be lost.

The style improvements are mostly abstractions.  We're not in some
sort of perfect world yet, but it seems toh current state is okayish
to work with. The templates have become very dirty by adding the
animation primitives.  There's a good chance we should create
abstractions on top of ember-animated or experiment with better
approaches to rendering such content.