Allow children to be created and add some styling
The new styling touches on the newly created functionality, but it
also allows for the children to be created in a nice way.  This should
be sufficient for managing the product groups.

The next challenge will be the management of products themselves, for
which we need to come up with a new interface.