Create subroute of product-groups to manage single product group
The position of a product-group can easily be managed in the tree of
product groups.  Managing the contents of the product group itself
could better be managed in a custom view.

This change introduces the needed nested route and moves logic that
was previously in product-groups route/controller/template into the
product-groups.index route/controller/template.