Allow prices to be calculated automatically
Not all logical elements of this flow have been tested, but it's
important as the price calculation is a rather messy portion of the
application.  The whole price model is quite complex and we haven't
thought about how we should present this to end-users leading to a
complex interface where hand-holding is very welcome.