Order offers by amount and put product save button higher up
The product save button was way downward.  It makes sense to put this
higher, closer to where people are editing when they need to press the
save button.  The amounts are now ordered by value (not by unit _and_
value, just by value).  This should be sufficient to order offers in
most cases.