Resolve linter warnings and fix computation bug
The automatic computation of the product prices was not correct.  Due
to some random browser display issue, we started tackling warnings
linter warnings in the hopes that would resolve the issue.  Committing
all of these in one go as it's fairly hard to split them and I doubt
we will need this separately.